requirements.txt
pytorch-ignite
torchsummary
wandb
#pandas
numpy
torch
torchvision
matplotlib
